% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_met_files_ler.R
\name{generate_met_files_ler}
\alias{generate_met_files_ler}
\title{Convert NOAA forecasts to GLM format}
\usage{
generate_met_files_ler(
  obs_met_file = NULL,
  out_dir,
  forecast_dir = NULL,
  config
)
}
\arguments{
\item{obs_met_file}{filepath; to observed met file}

\item{out_dir}{filepath; to directory for outputs}

\item{forecast_dir}{filepath; to directory with forecasts}

\item{config}{list; list of configurations}
}
\value{
vector; of filenames
}
\description{
Convert NOAA forecasts to GLM format
}
