% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_met_files_ler.R
\name{generate_met_files_ler}
\alias{generate_met_files_ler}
\title{Convert NOAA forecasts to GLM format}
\usage{
generate_met_files_ler(
  obs_met_file = NULL,
  out_dir,
  forecast_dir = NULL,
  config
)
}
\value{

}
\description{
Convert NOAA forecasts to GLM format
}
